web_manager/app/requirements.txt
WTForms==3.1.1
flask-wtf==1.2.1
Bootstrap-Flask==2.3.3